Save local changes

+++ CinemaProject_Avalonia/Models/MainWindowModel.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Text;
+using System.Text.Json;
 using System.Threading.Tasks;
 using Cinema.Dto;
 using CinemaProject.Dto;
@@ -20,7 +22,6 @@ namespace CinemaProject_Avalonia.Models
             _client = _session.Client;
         }
 
-
         // ===================== MOVIES =====================
 
         public async Task<List<MovieDto>> GetAllMovies()
@@ -35,19 +36,19 @@ namespace CinemaProject_Avalonia.Models
 
         public async Task NewMovie(NewMovieDto dto)
         {
-            var response = await _client.PostAsJsonAsync("/newmovie", dto);
+            var response = await _client.PostAsJsonAsync("api/admin/newmovie", dto);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task ModifyMovie(MovieDto dto, int movieId)
         {
-            var response = await _client.PutAsJsonAsync($"/modifymovie?movieId={movieId}", dto);
+            var response = await _client.PutAsJsonAsync($"api/admin/modifymovie?movieId={movieId}", dto);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task DeleteMovie(int movieId)
         {
-            var response = await _client.DeleteAsync($"/deletemovie?movieId={movieId}");
+            var response = await _client.DeleteAsync($"api/admin/deletemovie?movieId={movieId}");
             response.EnsureSuccessStatusCode();
         }
 
@@ -55,7 +56,16 @@ namespace CinemaProject_Avalonia.Models
 
         public async Task<List<FilmScreeningDto>> GetAllScreenings()
         {
-            return await _client.GetFromJsonAsync<List<FilmScreeningDto>>("api/cinema/getallscreenings");
+            var response = await _client.GetAsync("/getallscreenings");
+
+            Console.WriteLine(response.StatusCode);
+
+            var content = await response.Content.ReadAsStringAsync();
+            Console.WriteLine(content);
+
+            response.EnsureSuccessStatusCode();
+
+            return JsonSerializer.Deserialize<List<FilmScreeningDto>>(content);
         }
 
         public async Task<List<FilmScreeningDto>> GetScreeningDetails(DateTimeOffset time)
@@ -68,21 +78,23 @@ namespace CinemaProject_Avalonia.Models
             return await _client.GetFromJsonAsync<List<FilmScreeningDto>>("api/cinema/getupcomingscreenings");
         }
 
-        public async Task NewScreening(NewScreeningDto dto)
+        public async Task<NewScreeningDto> NewScreening(NewScreeningDto dto)
         {
-            var response = await _client.PostAsJsonAsync("/newscreening", dto);
+            var response = await _client.PostAsJsonAsync("api/admin/newscreening", dto);
             response.EnsureSuccessStatusCode();
+            var result = await response.Content.ReadFromJsonAsync<NewScreeningDto>();
+            return result!;
         }
 
         public async Task ModifyFilmScreening(FilmScreeningDto dto, int screeningId)
         {
-            var response = await _client.PutAsJsonAsync($"/modifyfilmscreening?screeningId={screeningId}", dto);
+            var response = await _client.PutAsJsonAsync($"api/admin/modifyfilmscreening?screeningId={screeningId}", dto);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task DeleteScreening(int screeningId)
         {
-            var response = await _client.DeleteAsync($"/deletescreening?screeningId={screeningId}");
+            var response = await _client.DeleteAsync($"api/admin/deletescreening?screeningId={screeningId}");
             response.EnsureSuccessStatusCode();
         }
 
@@ -90,12 +102,12 @@ namespace CinemaProject_Avalonia.Models
 
         public async Task<List<TicketDto>> GetAllTickets()
         {
-            return await _client.GetFromJsonAsync<List<TicketDto>>("api/cinema/getallticket");
+            return await _client.GetFromJsonAsync<List<TicketDto>>("getallticket");
         }
 
         public async Task ModifyTicket(TicketDto dto, int ticketId)
         {
-            var response = await _client.PutAsJsonAsync($"/modifyticket?ticketId={ticketId}", dto);
+            var response = await _client.PutAsJsonAsync($"api/admin/modifyticket?ticketId={ticketId}", dto);
             response.EnsureSuccessStatusCode();
         }
 
@@ -108,23 +120,23 @@ namespace CinemaProject_Avalonia.Models
 
         public async Task<List<UserDto>> GetAllUsers()
         {
-            return await _client.GetFromJsonAsync<List<UserDto>>("/getalluser");
+            return await _client.GetFromJsonAsync<List<UserDto>>("api/admin/getalluser");
         }
 
         public async Task<List<UserDto>> SearchUser(string item)
         {
-            return await _client.GetFromJsonAsync<List<UserDto>>($"/searchuser?item={item}");
+            return await _client.GetFromJsonAsync<List<UserDto>>($"api/admin/searchuser?item={item}");
         }
 
         public async Task DeleteUser(int userId)
         {
-            var response = await _client.DeleteAsync($"/deleteuser?userId={userId}");
+            var response = await _client.DeleteAsync($"api/admin/deleteuser?userId={userId}");
             response.EnsureSuccessStatusCode();
         }
 
         public async Task ChangeRole(int userId)
         {
-            var response = await _client.PutAsync($"/changerole?userId={userId}", null);
+            var response = await _client.PutAsync($"api/admin/changerole?userId={userId}", null);
             response.EnsureSuccessStatusCode();
         }
 
@@ -132,18 +144,18 @@ namespace CinemaProject_Avalonia.Models
 
         public async Task<List<PaymentReservationDto>> GetAllReservations()
         {
-            return await _client.GetFromJsonAsync<List<PaymentReservationDto>>("/getallreservation");
+            return await _client.GetFromJsonAsync<List<PaymentReservationDto>>("api/admin/getallreservation");
         }
 
         public async Task ModifyReservation(PaymentReservationDto dto, int reservationId)
         {
-            var response = await _client.PutAsJsonAsync($"/modifyreservation?reservationId={reservationId}", dto);
+            var response = await _client.PutAsJsonAsync($"api/admin/modifyreservation?reservationId={reservationId}", dto);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task DeleteReservation(int reservationId)
         {
-            var response = await _client.DeleteAsync($"/deletereservation?reservationId={reservationId}");
+            var response = await _client.DeleteAsync($"api/admin/deletereservation?reservationId={reservationId}");
             response.EnsureSuccessStatusCode();
         }
 
@@ -151,13 +163,13 @@ namespace CinemaProject_Avalonia.Models
 
         public async Task UploadImage(ImageDto dto)
         {
-            var response = await _client.PostAsJsonAsync("/uploadimage", dto);
+            var response = await _client.PostAsJsonAsync("api/admin/uploadimage", dto);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task DeleteImage(int imageId)
         {
-            var response = await _client.DeleteAsync($"/deleteimage?imageId={imageId}");
+            var response = await _client.DeleteAsync($"api/admin/deleteimage?imageId={imageId}");
             response.EnsureSuccessStatusCode();
         }
